huh
ok archaicfix *was* compatible but i never actually checked the archaicfix source for potential incompats... suffice to say, its not compatible.

all changes here are just spotless changes.

+++ src/main/java/com/gamma/spool/core/SpoolCompat.java
@@ -176,8 +176,7 @@ public class SpoolCompat {
             outputFile = new File(Launch.minecraftHome, "spool/spool-compat.log");
             outputFile.getParentFile()
                 .mkdirs();
-            if(outputFile.exists())
-                outputFile.delete();
+            if (outputFile.exists()) outputFile.delete();
             try {
                 outputFileWriter = new FileWriter(outputFile);
             } catch (IOException e) {
